requirements.txt
Django==4.2.7
# django-cors-headers==4.3.0
djangorestframework==3.14.0
# Pillow==10.1.0